Split into multiple files (#51)
* Increment version: 1.7.8 -> 1.7.9
* Split into multiple files
