+++ setup.py
@@ -9,7 +9,7 @@ with open("README.md", "r", encoding="utf-8") as fh:
 
 __metadata__ = \
 {
-    "version" :  "1.7.8",
+    "version" :  "1.7.9",
     "requirements" :  ["aiohttp>=3.6.0,<3.9.0", "PyNaCl"],
     "minimum_py_version" :  "3.8",
 }
