+++ src/framework/__init__.py
@@ -2,6 +2,13 @@
     DISCORD ADVERTISEMENT FRAMEWORK (DAF)
     Author      :: David Hozic
     Copyright   :: Copyright (c) 2022 David Hozic
-    Version     :: V1.7.8
+    Version     :: V1.7.9
 """
-from .__core__ import *
+import _discord as discord
+from .client import *
+from .const import *
+from .core import *
+from .dtypes import *
+from .guild import *
+from .message import *
+from .tracing import *
