+++ src/framework/const.py
@@ -1,0 +1,12 @@
+#######################################################################
+# Contants
+#######################################################################
+C_DAY_TO_SECOND = 86400
+C_HOUR_TO_SECOND= 3600
+C_MINUTE_TO_SECOND = 60
+
+C_RT_AVOID_DELAY     = 1    # Rate limit avoidance delay
+C_TASK_SLEEP_DELAY   = 0.1  # Advertiser task sleep
+C_VC_CONNECT_TIMEOUT = 1    # Timeout of voice channels
+
+C_FILE_NAME_FORBIDDEN_CHAR = ('<','>','"','/','\\','|','?','*',":")
